+++ include/dt-bindings/iio/adc/adi,ad4695.h
